Remove the dependency on core_io.
Implement a subset of the std::io::{Error,ErrorKind}
types and std::io::{Read,Write} traits.  Only enough
for nanobl-rs.

These are public so that nanobl-rs can implement them.
At some point, we may wish to extract them into e.g.,
`nanoio`, but this is sufficient for the time being.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>